requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
pycparser==2.21
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
typing_extensions==4.3.0
